MR-385: componentization of img persistance - functional in member with session storage manage
