+++ src/Components/Users/Admin/profile/index.js
@@ -1,12 +1,11 @@
 import React, { useEffect, useState } from 'react';
 import styles from './profile.module.css';
-import { ButtonForm, ProfilePicList } from 'Components/Shared';
+import { ButtonForm } from 'Components/Shared';
 import { useSelector, useDispatch } from 'react-redux';
 import { useHistory } from 'react-router-dom';
 import { getAllAdmins } from 'redux/admins/thunks';
 import { getFirebaseUidFromToken } from 'helper/firebase';
 import 'firebase/compat/auth';
-import ImageUploader from 'Components/Shared/ChangeImg/imageUploader';
 // const editProfilePicBtn = `${process.env.PUBLIC_URL}/assets/images/image.png`;
 
 const AdminProfile = ({ testId }) => {
@@ -16,8 +15,6 @@ const AdminProfile = ({ testId }) => {
   const admins = useSelector((state) => state.admins.list);
   const admin = admins.find((oneAdmin) => oneAdmin.email === userCurrent);
   const [profilePic, setProfilePic] = useState('');
-  const [photoEdit, setPhotoEdit] = useState(false);
-  // const [counter, setCounter] = useState(0);
 
   const defaultProfile = !profilePic ? (
     <div className={styles.defaultImg}>
@@ -54,10 +51,7 @@ const AdminProfile = ({ testId }) => {
   if (!admin) {
     return null;
   }
-  // const togglePhotoEdit = () => {
-  //   setCounter(counter + 1);
-  //   setPhotoEdit(!photoEdit);
-  // };
+
   return (
     <div className={styles.wholeContainer}>
       <section className={styles.container} data-testid={testId}>
@@ -67,13 +61,6 @@ const AdminProfile = ({ testId }) => {
           ) : (
             profilePic
           )}
-          <ImageUploader />
-          <ProfilePicList
-            profilePic={setProfilePic}
-            photoEdit={setPhotoEdit}
-            show={photoEdit}
-            // counter={counter}
-          />
           <h1 className={styles.adminName}>
             {admin.firstName} {admin.lastName}
           </h1>
