+++ src/Components/Users/Trainer/profile/index.js
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import styles from './profile.module.css';
-import { ButtonForm, ProfilePicList } from 'Components/Shared';
+import { ButtonForm } from 'Components/Shared';
 import { useSelector, useDispatch } from 'react-redux';
 import { useHistory } from 'react-router-dom';
 import { getTrainers } from 'redux/trainers/thunks';
@@ -71,12 +71,6 @@ const TrainerProfile = () => {
             src={editProfilePicBtn}
             onClick={togglePhotoEdit}
             alt="camera"
-          />
-          <ProfilePicList
-            profilePic={setProfilePic}
-            photoEdit={setPhotoEdit}
-            show={photoEdit}
-            counter={counter}
           />
           <h1 className={styles.adminName}>
             {trainer.firstName} {trainer.lastName}
